+++ src/effects/StatusEffect.ts
