+++ src/effects/StatusEffectConfig.hbs
