+++ src/effects/StatusEffectConfig.ts
