+++ src/effects/StatusEffectScheduler.ts
